backend/requirements.txt
Flask>=3.0
Flask-SQLAlchemy>=3.1
Flask-Migrate>=4.0
Flask-CORS>=4.0
cryptography>=42.0
python-dotenv>=1.0
gunicorn>=22.0
